created cusom components package for better ordering of the problem.
