+++ inventarios-desktop-testing-api/src/test/java/desktop/GUITest.java
@@ -32,9 +32,9 @@ import inventarios.desktop.pageobjects.LoginWindowPageObject;
 import inventarios.desktop.pageobjects.MainMenuPageObject;
 import inventarios.desktop.pageobjects.OrderManagementPageObject;
 import inventarios.desktop.navigation.NavigationHandler;
-import inventarios.service.LoginUsersService;
-import inventarios.service.OrderService;
-import inventarios.service.ProductService;
+import inventarios.service.restclient.LoginUsersService;
+import inventarios.service.restclient.OrderService;
+import inventarios.service.restclient.ProductService;
 import inventarios.to.LoginUser;
 import inventarios.util.FontFactory;
 import inventarios.util.ShutdownManager;
