+++ inventarios-desktop-testing-api/src/test/java/desktop/StepDefinition.java
@@ -33,8 +33,8 @@ import inventarios.desktop.pageobjects.InventoryManagementPageObject;
 import inventarios.desktop.pageobjects.LoginWindowPageObject;
 import inventarios.desktop.pageobjects.MainMenuPageObject;
 import inventarios.desktop.pageobjects.OrderManagementPageObject;
-import inventarios.service.LoginUsersService;
-import inventarios.service.ProductService;
+import inventarios.service.restclient.LoginUsersService;
+import inventarios.service.restclient.ProductService;
 import inventarios.to.LoginUser;
 import inventarios.util.FontFactory;
 import inventarios.util.ShutdownManager;
