+++ inventarios-desktop/src/main/java/inventarios/desktop/EmployeeRegistration.java
@@ -17,14 +17,13 @@
 package inventarios.desktop;
 
 import inventarios.desktop.navigation.NavigationHandler;
-import inventarios.service.EmployeeService;
+import inventarios.service.restclient.EmployeeService;
 import inventarios.to.EmployeeDetail;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.stereotype.Component;
 
 import javax.swing.BorderFactory;
-import javax.swing.GroupLayout;
 import javax.swing.ImageIcon;
 import javax.swing.JButton;
 import javax.swing.JFrame;
