+++ inventarios-desktop/src/main/java/inventarios/desktop/ShoppingWindow.form
@@ -208,8 +208,19 @@
               </AuxValues>
             </Component>
             <Component class="javax.swing.JComboBox" name="cmbProvider">
+              <Properties>
+                <Property name="editable" type="boolean" value="true"/>
+                <Property name="editor" type="javax.swing.ComboBoxEditor" editor="org.netbeans.modules.form.RADConnectionPropertyEditor">
+                  <Connection code="new ProviderItemEditor()" type="code"/>
+                </Property>
+                <Property name="renderer" type="javax.swing.ListCellRenderer" editor="org.netbeans.modules.form.RADConnectionPropertyEditor">
+                  <Connection code="new ProviderItemRenderer()" type="code"/>
+                </Property>
+              </Properties>
               <AuxValues>
-                <AuxValue name="JavaCodeGenerator_TypeParameters" type="java.lang.String" value="&lt;String&gt;"/>
+                <AuxValue name="JavaCodeGenerator_TypeParameters" type="java.lang.String" value="&lt;Provider&gt;"/>
+                <AuxValue name="JavaCodeGenerator_VariableLocal" type="java.lang.Boolean" value="false"/>
+                <AuxValue name="JavaCodeGenerator_VariableModifier" type="java.lang.Integer" value="2"/>
               </AuxValues>
             </Component>
             <Component class="javax.swing.JTextField" name="txtAdress">
