+++ inventarios-restservice/src/main/java/inventarios/controller/UsersController.java
@@ -18,12 +18,13 @@ package inventarios.controller;
 
 import inventarios.service.LoginUsersService;
 import inventarios.to.LoginUser;
-import java.util.logging.Level;
 import lombok.extern.java.Log;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
+
+import java.util.logging.Level;
 
 /**
  *
